# Conflicts:
#	src/debug.c

+++ src/util.c
@@ -261,7 +261,7 @@ u32 CalcByteArraySum(const u8 *data, u32 length)
     return sum;
 }
 
-void BlendPalette(u16 palOffset, u16 numEntries, u8 coeff, u16 blendColor)
+void BlendPalette(u16 palOffset, u16 numEntries, u8 coeff, u32 blendColor)
 {
     u16 i;
     for (i = 0; i < numEntries; i++)
